Refactoring: GridCreatingConditionTriangle::DefinePolygonNewPointCommand implementation is separated from gridcreatingconditiontriangle.cpp.
